# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Streaming service YouTube alike",
        "title": "ROSS API",
        "contact": {
            "name": "Ro Ngoc Vo",
            "url": "github.com/rovn208",
            "email": "ngocro208@gmail.com"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/videos": {
            "get": {
                "description": "Get list video",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Get list video",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.videoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create new video",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Create new video",
                "parameters": [
                    {
                        "minLength": 6,
                        "description": "Video title",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "foldername/video.mp4",
                        "name": "stream_url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Video description",
                        "name": "description",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "https://i.ytimg.com/vi/-uFQzcY7YHc/hqdefault.jpg?sqp=-oaymwEbCKgBEF5IVfKriqkDDggBFQAAiEIYAXABwAEG\u0026rs=AOn4CLBdAOc5E4H_G09C5wqorhYRsUwQrQ",
                        "name": "thumbnail_url",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "123451",
                        "name": "created_by",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.videoResponse"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/videos/youtube": {
            "post": {
                "description": "Add video via youtube video url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Add video via youtube video url",
                "parameters": [
                    {
                        "description": "https://www.youtube.com/watch?v=-uFQzcY7YHc",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "created video successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/videos/{id}": {
            "get": {
                "description": "Get video",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Get video",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "12345",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.videoResponse"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update video",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Update video",
                "parameters": [
                    {
                        "minLength": 6,
                        "description": "Video title",
                        "name": "title",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "foldername/video.mp4",
                        "name": "stream_url",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Video description",
                        "name": "description",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "https://i.ytimg.com/vi/-uFQzcY7YHc/hqdefault.jpg?sqp=-oaymwEbCKgBEF5IVfKriqkDDggBFQAAiEIYAXABwAEG\u0026rs=AOn4CLBdAOc5E4H_G09C5wqorhYRsUwQrQ",
                        "name": "thumbnail_url",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.videoResponse"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete video",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Delete video",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "12345",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted video successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"error\": \"error message\"}",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.videoResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "stream_url": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}